+++ apps/frontend/app/(website)/studio/studioMetadata.tsx
@@ -1,0 +1,26 @@
+export const metadata = {
+  title: "Codemod Studio",
+  description:
+    "Build codemods instantly with the help of AI, specialized helpers, debuggers, and a vibrant community of Codemod Champions.",
+  keywords:
+    "codemod studio, codemod, codemods, codebase migration, code evolution",
+  manifest: "./manifest.json",
+  icons: {
+    icon: "./favicon.ico",
+    apple: "./logo192.png",
+  },
+  twitter: {
+    title: "Codemod Studio",
+    description:
+      "Build codemods instantly with the help of AI, specialized helpers, debuggers, and a vibrant community of Codemod Champions.",
+    creator: "@codemod",
+  },
+  openGraph: {
+    title: "Codemod Studio",
+    description:
+      "Build codemods instantly with the help of AI, specialized helpers, debuggers, and a vibrant community of Codemod Champions.",
+    url: "https://codemod.studio",
+    siteName: "Codemod Studio",
+    locale: "en_US",
+  },
+};
