+++ apps/frontend/app/api/newsletter-form/route.ts
@@ -1,6 +1,7 @@
 import { env } from "@/env";
 import { headers } from "next/headers";
 import { type NextRequest, NextResponse } from "next/server";
+
 /**
  * Handles form submissions from Form.tsx by ensuring it isn't spam and sending the data to the coresponding Notion database.
  *
